+++ src/ActionBar/ActionBar.test.tsx
@@ -10,6 +10,8 @@ describe('ActionBar', () => {
         onFocus: jest.fn(),
         onDelete: jest.fn(),
         onEnableVectorMode: jest.fn(),
+        onExport: jest.fn(),
+        onImport: jest.fn(),
         deleteInactive: true,
         editable: false,
         isVectorModeEnabled: false
@@ -20,7 +22,6 @@ describe('ActionBar', () => {
     });
 
     describe('WHEN editable is falsy', () => {
-
         it('should only render the action buttons', () => {
             const wrapper = render(<ActionBar {...props} />);
             expect(wrapper.getByText(LABELS.FOCUS)).toBeTruthy();
@@ -37,7 +38,6 @@ describe('ActionBar', () => {
     });
 
     describe('WHEN editable is truthy', () => {
-
         it('should render the actions', () => {
             const wrapper = render(<ActionBar {...props} editable />);
             expect(wrapper.getByText(LABELS.FOCUS)).toBeTruthy();