+++ src/ActionBar/ActionBar.tsx
@@ -30,11 +30,23 @@ export interface Props {
 
     onDelete: () => void;
     deleteInactive: boolean;
+
+    onExport: () => void;
+    onImport: () => void;
 }
 
-export const ActionBar: FunctionComponent<Props> = ({ editable, deleteInactive, isVectorModeEnabled, onEnableVectorMode, onFocus, onDelete }) => (
+export const ActionBar: FunctionComponent<Props> = ({
+    editable,
+    deleteInactive,
+    isVectorModeEnabled,
+    onEnableVectorMode,
+    onFocus,
+    onDelete,
+    onExport,
+    onImport
+}) => (
     <Container>
-        {editable &&
+        {editable && (
             <>
                 <ActionButton
                     onClick={onEnableVectorMode}
@@ -53,10 +65,15 @@ export const ActionBar: FunctionComponent<Props> = ({ editable, deleteInactive, 
                     {LABELS.DELETE}
                 </ActionButton>
             </>
-        }
+        )}
+        <ActionButton onClick={onImport} icon={ActionButtonIcons.IMPORT}>
+            {LABELS.IMPORT}
+        </ActionButton>
+        <ActionButton onClick={onExport} icon={ActionButtonIcons.EXPORT}>
+            {LABELS.EXPORT}
+        </ActionButton>
         <ActionButton onClick={onFocus} icon={ActionButtonIcons.FRAME}>
             {LABELS.FOCUS}
         </ActionButton>
     </Container>
 );
-