+++ src/PolygonDraw/PolygonVertex.tsx
@@ -1,7 +1,7 @@
 import 'leaflet-path-drag';
-import { CircleMarker as CircleMarkerElement, LatLng } from 'leaflet';
 import React from 'react';
-import { CircleMarker } from 'react-leaflet';
+import { CircleMarker, LatLng } from 'leaflet';
+import { CircleMarker as ReactLeafletCircleMarker } from 'react-leaflet';
 
 import { Coordinate } from 'types';
 import { createLeafletLatLngFromCoordinate } from '../helpers';
@@ -26,7 +26,7 @@ export interface State {
 }
 
 export class PolygonVertex extends React.Component<Props, State> {
-    private circleMarkerElement: CircleMarkerElement;
+    private circleMarkerElement: CircleMarker;
     state = {
         isHovered: false,
         isDragged: false,
@@ -34,7 +34,7 @@ export class PolygonVertex extends React.Component<Props, State> {
     };
 
     static getDerivedStateFromProps = (props: Props, state: State) => {
-        return state.isDragged ? state : { ...state, latLng: createLeafletLatLngFromCoordinate(props.coordinate)};
+        return state.isDragged ? state : { ...state, latLng: createLeafletLatLngFromCoordinate(props.coordinate) };
     };
 
     componentDidMount() {
@@ -68,11 +68,13 @@ export class PolygonVertex extends React.Component<Props, State> {
         this.props.onDragStart(event.target.getLatLng(), this.props.index);
         this.setState({ isDragged: true });
     };
+
     handleDrag = (event: DragEvent) => {
         if (this.state.isDragged) {
             this.props.onDrag(event.target.getLatLng());
         }
     };
+
     handleDragEnd = () => {
         this.props.onDragEnd();
         this.setState({ isDragged: false });
@@ -85,11 +87,11 @@ export class PolygonVertex extends React.Component<Props, State> {
         const hoveredAndNotSelected = isHovered && !isSelected;
 
         return (
-            <CircleMarker
+            <ReactLeafletCircleMarker
                 ref={this.setCircleMarkerRef}
                 fillColor={MAP.VERTEX_FILL_COLOR}
                 fillOpacity={1}
-                color={MAP.POLYGON_COLOR}
+                color={MAP.POLYGON_ACTIVE_COLOR}
                 opacity={hoveredAndNotSelected ? 0.6 : 1}
                 weight={selectedOrHovered ? 4 : 1}
                 radius={selectedOrHovered ? 6 : 4}