+++ src/PolygonDraw/validators.ts
@@ -1,8 +1,5 @@
 import { Coordinate } from 'types';
 import { isPolygonClosed } from '../helpers';
 
-export const isValidPolygon = (polygon: Coordinate[]) => (
-    polygon &&
-    polygon.length > 3 &&
-    isPolygonClosed(polygon)
-);
+export const isValidPolygon = (coordinates: Coordinate[]): boolean =>
+    !!coordinates && coordinates.length > 3 && isPolygonClosed(coordinates);