+++ src/common/colors.tsx
@@ -1,5 +1,13 @@
 export const ACTION_BLUE_900 = '#096bdb';
+export const AUTHENTIC_BLUE_1100 = '#000f1f';
 export const AUTHENTIC_BLUE_900 = '#001e3e';
+export const AUTHENTIC_BLUE_550 = '#637689';
+export const AUTHENTIC_BLUE_350 = '#9CA7B4';
 export const AUTHENTIC_BLUE_200 = '#c6cdd4';
+export const AUTHENTIC_BLUE_50 = '#f1f7fd';
 export const FREEDOM_RED_900 = '#ff0a2b';
+export const NEGATIVE_ORANGE_50 = '#fff6f1';
+export const NEGATIVE_ORANGE_900 = '#ff5e05';
+export const POSITIVE_GREEN_50 = '#f1faf5';
+export const POSITIVE_GREEN_900 = '#069d4f';
 export const WHITE = '#ffffff';