+++ CHANGELOG.md
@@ -1,13 +1,28 @@
 # Changelog
 
-## [~] 
+## [1.1.0.rc3] - 12.03.2020
+
+-   Fix: Clear selection after import
+-   Fix: Focus on polygon after import
+-   Fix: Prevent default behaviour for keyevents
+
+## [1.1.0.rc2] - 11.03.2020
+
+-   Feat: Import polygons from GeoJSON
+
+## [1.1.0.rc1] - 06.03.2020
+
+-   Feat: Rectangle selection
+-   Feat: Export polygons as GeoJSON, JTS or LatLng
 
 ## [0.0.1] - 08.03.2019
-- Fix: fix leaflet-path-drag dependency
-- Fix: position action bar at the right bottom of the map
+
+-   Fix: fix leaflet-path-drag dependency
+-   Fix: change action bar position to the right bottom of the map
 
 ## [0.0.0] - 08.03.2019
-- Add tests
-- Add storybook
-- Add PoylgonDraw component
-- Create base project
+
+-   Add tests
+-   Add storybook
+-   Add PoylgonDraw component
+-   Create base project